fix: correct shift logic and lending CEI pattern
  - Fix skim ratio application in preShiftPositions (apply to discovery, not anchor)
  - Refresh sqrtRatioX96 before preShiftPositions to use latest price
  - Simplify fixInbalance call to use 1 wei placeholder
  - Fix CEI pattern in LendingVault.extendLoan (state updates after transfers)
  - Widen PresaleProtocolParams contribution ratios to uint32 for PPM support
  - Add PPM constant and applyPpm utility function
  - Consolidate error imports to types/Errors.sol

+++ src/types/Types.sol
@@ -186,8 +186,8 @@ struct LivePresaleParams {
 /// @dev Only *_Bps fields are scaled by 10_000 (bps). Others remain plain integers.
 struct PresaleProtocolParams {
     uint256 maxSoftCap;                 // unchanged
-    uint16  minContributionRatioBps;    // NEW: bps (0–10_000)
-    uint16  maxContributionRatioBps;    // NEW: bps (0–10_000)
+    uint32  minContributionRatioBps;    // NEW: bps (0–10_000)
+    uint32  maxContributionRatioBps;    // NEW: bps (0–10_000)
     uint256 presalePercentage;          // unchanged (integer percent)
     uint256 minDuration;                // unchanged (seconds)
     uint256 maxDuration;                // unchanged (seconds)
